feat(examples): add typescript http hello-world actor
As Javascript support for WebAssembly progresses along with the `jco`
toolchain, it's possible to build actors that work with the wasmcloud
ecosystem, leveraging the existing host, provider and actor ecosystem.

This commit adds a Typescript HTTP hello-world actor which shows how to use
the Typescript and `jco` ecosystems together with wasmcloud.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>